+++ changelog.md
@@ -1,3 +1,5 @@
+# v0.1.6-beta.1
+- Fixed the popup to fit based on your screen resolution
 # v0.1.5-beta.1
 - Added **Sound Effect** & **Stop All Sounds** Events
 - Changed the **Wait** Action value to use float value instead of just integer.